refactor : modify rmserver packages' structure
