+++ rmserver/http/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "rmvserver_http",
+  "version": "1.0.0",
+  "description": "",
+  "main": "src/main/app.ts",
+  "scripts": {
+    "compile": "npx tsc",
+    "start": "node dist/app.js",
+    "dev": "nodemon --watch \"*.ts\" --exec \"ts-node\" src/main/app.ts",
+    "server_build": "concurrently \"npx tsc --build .\" \"cd ../ros && colcon build --symlink-install\"",
+    "client_build": "cd ../../rmviz && npm run build",
+    "build": "sudo chmod +x ./script/build.sh && ./script/build.sh",
+    "launch": "sudo chmod +x ./script/launch.sh && ./script/launch.sh"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "body-parser": "^1.20.0",
+    "concurrently": "^7.2.2",
+    "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
+    "debug": "~2.6.9",
+    "express": "^4.19.2",
+    "morgan": "^1.10.0"
+  },
+  "devDependencies": {
+    "@types/cookie-parser": "^1.4.3",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.21",
+    "@types/http-errors": "^1.8.2",
+    "@types/morgan": "^1.9.9",
+    "@types/node": "^20.12.8",
+    "@types/socket.io": "^3.0.2",
+    "@types/typescript": "^2.0.0",
+    "concurrently": "^8.2.2",
+    "nodemon": "^3.1.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
+  }
+}
