+++ rmserver/http/script/launch.sh
