--- rmserver/ros/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/rmserver_kec
-[install]
-install_scripts=$base/lib/rmserver_kec
