+++ rmserver/rqtt/config/path_config.yaml
@@ -1,0 +1,5 @@
+/rmserver_rqtt:
+  ros__parameters:
+    current_mqtt_file: "RobotData/mqtt/mqtt.json"
+    current_map_config_file: "RobotData/maps/kecd_path/config/config.ini"
+    current_urdf_file: "RobotData/urdf/kec_model.urdf"
