+++ rmserver/rqtt/json/detected_object.json
