+++ rmserver/rqtt/package.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>rmserver_rqtt</name>
+  <version>0.1.0</version>
+  <description>TODO: Package description</description>
+  <maintainer email="naru5135@wavem.net">reidlo</maintainer>
+  <license>TODO: License declaration</license>
+
+  <depend>rclpy</depend>
+  <depend>ktp_data_msgs</depend>
+  <depend>obstacle_msgs</depend>
+  <depend>can_msgs</depend>
+  <depend>route_msgs</depend>
+
+  <test_depend>ament_copyright</test_depend>
+  <test_depend>ament_flake8</test_depend>
+  <test_depend>ament_pep257</test_depend>
+  <test_depend>python3-pytest</test_depend>
+
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
+</package>
