+++ rmserver/rqtt/rmserver_rqtt/application/request/obstacle.py
