+++ rmserver/rqtt/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/rmserver_rqtt
+[install]
+install_scripts=$base/lib/rmserver_rqtt
